+++ installer/Build-Installer.ps1
@@ -1,7 +1,7 @@
 ﻿[CmdletBinding()]
 param(
     [ValidatePattern("^\d+\.\d+\.\d+$")]
-    [string]$Version = "1.9.5"
+    [string]$Version = "1.9.6"
 )
 
 Set-StrictMode -Version 2.0
@@ -133,4 +133,5 @@ using System.Reflection;
         }
     }
 }
+